requirements.txt
numpy
requests
pandas
Pillow